# Conflicts:
#	crates/service/job/inference_job/src/job/job_types/mod.rs
#	crates/service/web/storyteller_web/src/http_server/endpoints/media_files/upload/drain_multipart_request.rs
#	crates/service/web/storyteller_web/src/http_server/endpoints/weights/mod.rs
#	crates/service/web/storyteller_web/src/http_server/endpoints/weights/upload_weight.rs
#	crates/service/web/storyteller_web/src/routes.rs

+++ crates/service/web/storyteller_web/src/http_server/common_responses/pagination_page.rs
@@ -1,3 +1,5 @@
+use utoipa::ToSchema;
+
 /// Pagination by page id
 /// This type of pagination is by "page id" and "page count".
 /// This should never be used to walk the entire public database as competitors and investors
@@ -5,13 +7,13 @@
 /// This is best used for user profiles and scoped down results.
 ///
 /// See `PaginationCursors` for the other type of pagination.
-#[derive(Serialize)]
+#[derive(Serialize, ToSchema)]
 pub struct PaginationPage {
   /// The current page number
-  pub current: u64,
+  pub current: usize,
 
   /// How many pages there are in total
-  pub total_page_count: u64,
+  pub total_page_count: usize,
 }
 
 #[cfg(test)]
